+++ easyeditor/editors/batch_editor.py
@@ -14,6 +14,7 @@ class BatchEditor(Enum):
     DPO = "DPO"
     EMMET = "EMMET"
     ALPHAEDIT = "AlphaEdit"
+    ULTRAEDIT = "ULTRAEDIT"
 
     @staticmethod
     def is_batchable_method(alg_name: str):
@@ -27,4 +28,5 @@ class BatchEditor(Enum):
             or alg_name == BatchEditor.QLoRA.value \
             or alg_name == BatchEditor.LoRA.value \
             or alg_name == BatchEditor.DPO.value \
+            or alg_name == BatchEditor.ULTRAEDIT.value \
             or alg_name == BatchEditor.EMMET.value or alg_name == BatchEditor.ALPHAEDIT.value
