+++ easyeditor/models/__init__.py
@@ -20,3 +20,4 @@ from .deco import *
 from .dola import *
 from .deepedit_api import *
 from .defer import *
+from .ultraedit import *
