+++ easyeditor/trainer/algs/__init__.py
@@ -2,3 +2,4 @@ from .editable_model import *
 from .MEND import *
 from .SERAC import *
 from .MALMEN import *
+from .ULTRAEDIT import *
