+++ examples/run_ultraedit_editing.sh
@@ -9,14 +9,25 @@
 # unset CUDA_VISIBLE_DEVICES
 
 
+# python run_ultraedit_editing.py \
+#     --editing_method=UltraEdit \
+#     --hparams_dir=../hparams/ULTRAEDIT/gpt-j-6B.yaml \
+#     --data_dir=../data/ultraedit \
+#     --ds_size=20000 \
+#     --batch_size=100 \
+#     --data_type=zsre \
+#     --sequential_edit
+
 python run_ultraedit_editing.py \
     --editing_method=UltraEdit \
-    --hparams_dir=../hparams/ULTRAEDIT/gpt-j-6B.yaml \
+    --hparams_dir=../hparams/ULTRAEDIT/llama3-8b.yaml \
     --data_dir=../data/ultraedit \
     --ds_size=20000 \
     --batch_size=100 \
     --data_type=zsre \
     --sequential_edit
+
+
 
 # python run_ultraedit_editing.py \
 #     --editing_method=UltraEdit \
