+++ hparams/ULTRAEDIT/phi-4.yaml
@@ -1,0 +1,37 @@
+alg_name: "ULTRAEDIT"
+device: 0
+
+model_name: your/path/phi-4
+class_name: AutoModelForCausalLM
+tokenizer_class: AutoTokenizer
+tokenizer_name: your/path/phi-4
+inner_params:
+# wikibigedit ultraeditbench
+- model.layers.30.mlp.down_proj
+- model.layers.31.mlp.down_proj
+- model.layers.32.mlp.down_proj
+- model.layers.33.mlp.down_proj
+- model.layers.34.mlp.down_proj
+- model.layers.35.mlp.down_proj
+- model.layers.36.mlp.down_proj
+- model.layers.37.mlp.down_proj
+- model.layers.38.mlp.down_proj
+
+# Method
+alg: UltraEdit
+dropout: 0.0
+no_grad_layers: null
+
+lr: 1e-6
+token: mask
+
+batch_size: 100
+batch_size_once: 10
+editor_batch_size: 1024
+silent: False
+half: true
+
+model_parallel: false
+
+# Output
+results_dir: ./results
