--- hparams/UltraEdit/gemma-3-27b.yaml
@@ -1,25 +1,0 @@
-name_or_path: your/path/gemma-3-27b-it
-name: gemma-3-27b
-model_vec_size: 17920
-class_name: AutoModelForCausalLM
-edit_modules:
-#wikibigedit ultraeditbench
-- language_model.model.layers.52.mlp.gate_proj
-- language_model.model.layers.53.mlp.gate_proj
-- language_model.model.layers.54.mlp.gate_proj
-- language_model.model.layers.55.mlp.gate_proj
-- language_model.model.layers.56.mlp.gate_proj
-- language_model.model.layers.57.mlp.gate_proj
-- language_model.model.layers.58.mlp.gate_proj
-- language_model.model.layers.59.mlp.gate_proj
-- language_model.model.layers.60.mlp.gate_proj
-- language_model.model.layers.52.mlp.up_proj
-- language_model.model.layers.53.mlp.up_proj
-- language_model.model.layers.54.mlp.up_proj
-- language_model.model.layers.55.mlp.up_proj
-- language_model.model.layers.56.mlp.up_proj
-- language_model.model.layers.57.mlp.up_proj
-- language_model.model.layers.58.mlp.up_proj
-- language_model.model.layers.59.mlp.up_proj
-- language_model.model.layers.60.mlp.up_proj
-half: True
