--- hparams/UltraEdit/llama-3-instruct.yaml
@@ -1,40 +1,0 @@
-name_or_path: ypur/path/Llama-3-8b
-name: llama-3-instruct
-model_vec_size: 18432
-class_name: AutoModelForCausalLM
-bit: 0
-edit_modules:
-# zsre wiki ultraeditbench
-- model.layers.11.mlp.gate_proj
-- model.layers.12.mlp.gate_proj
-- model.layers.13.mlp.gate_proj
-- model.layers.14.mlp.gate_proj
-- model.layers.15.mlp.gate_proj
-- model.layers.18.mlp.up_proj
-- model.layers.19.mlp.up_proj
-- model.layers.20.mlp.up_proj
-- model.layers.21.mlp.up_proj
-- model.layers.22.mlp.up_proj
-- model.layers.23.mlp.up_proj
-- model.layers.24.mlp.up_proj
-# fever
-# - model.layers.22.mlp.gate_proj
-# - model.layers.23.mlp.gate_proj
-# - model.layers.24.mlp.gate_proj
-# - model.layers.25.mlp.gate_proj
-# - model.layers.26.mlp.gate_proj
-# - model.layers.27.mlp.gate_proj
-# - model.layers.28.mlp.gate_proj
-# - model.layers.29.mlp.gate_proj
-# - model.layers.30.mlp.gate_proj
-# - model.layers.22.mlp.up_proj
-# - model.layers.23.mlp.up_proj
-# - model.layers.24.mlp.up_proj
-# - model.layers.25.mlp.up_proj
-# - model.layers.26.mlp.up_proj
-# - model.layers.27.mlp.up_proj
-# - model.layers.28.mlp.up_proj
-# - model.layers.29.mlp.up_proj
-# - model.layers.30.mlp.up_proj
-
-half: True
